[kernel] Rename namespace
What I did:
- rename "namespace op" to "namespace layer"
